pyproject.toml
[project]
name = "agentic-api-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "google-cloud-aiplatform>=1.128.0",
    "google-cloud-storage>=3.6.0",
    "google-genai>=1.52.0",
]
